refactor: keep bindings at gen/go/solidity-abi
Revert the directory move from the previous two commits: the
`gen/go/solidity-abi` location is mandated by an internal ADR, so the module
path, import paths, `replace` directive, Make targets, license config, CI path
filters, and Dependabot entry all go back to `gen/go/solidity-abi`.

The generation script stays colocated with the module it writes into, now at
`gen/go/solidity-abi/scripts/`, with `repo_root` adjusted for the extra depth.

The rest of the review feedback is unaffected and still applied: no committed
`go.work`, go-ethereum bumped to v1.17.5, and the script moved out of `e2e/`.
