+++ .licenserc.yaml
@@ -31,7 +31,7 @@ header:
     - "link/internal/tests/mocks/**"
     - "link/internal/service/**/*.mock.go"
     - "link/internal/chains/evm/contracts/**"
-    - "solidity-abi/*/contract.go"
+    - "gen/go/solidity-abi/*/contract.go"
 
     # Third-party and test Solidity sources retain their own SPDX identifiers.
     - "e2e/internal/harness/environment/solidityibc/contracts/src/**"
