+++ e2e/go.mod
@@ -290,8 +290,8 @@ require (
 	github.com/containerd/errdefs/pkg v0.3.0 // indirect
 	github.com/containerd/log v0.1.0 // indirect
 	github.com/containerd/platforms v0.2.1 // indirect
+	github.com/cosmos/ibc/gen/go/solidity-abi v0.0.0
 	github.com/cosmos/ibc/link v0.0.0
-	github.com/cosmos/ibc/solidity-abi v0.0.0
 	github.com/cosmos/solidity-ibc-eureka/packages/go-abigen v0.0.0-20260810020832-a40957eaf878
 	github.com/cpuguy83/dockercfg v0.3.2 // indirect
 	github.com/crate-crypto/go-eth-kzg v1.5.0 // indirect
@@ -348,6 +348,6 @@ require (
 )
 
 replace (
+	github.com/cosmos/ibc/gen/go/solidity-abi => ../gen/go/solidity-abi
 	github.com/cosmos/ibc/link => ../link
-	github.com/cosmos/ibc/solidity-abi => ../solidity-abi
 )
