+++ gen/go/solidity-abi/counter/contract.go
