+++ gen/go/solidity-abi/go.mod
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
-module github.com/cosmos/ibc/solidity-abi
+module github.com/cosmos/ibc/gen/go/solidity-abi
 
 go 1.26.4
