+++ gen/go/solidity-abi/testerc20/contract.go
